Modify id type int to Long
id의 타입을 int에서 Long으로 전체적인 수정

+++ app/src/main/java/bitcamp/carrot_thunder/dto/PostListResponseDto.java
@@ -17,7 +17,7 @@ import java.sql.Timestamp;
 public class PostListResponseDto {
 
   private User user;
-  private int postid;
+  private Long postid;
   private String title;
   private int viewCount;
   private Timestamp createdAt;
